--- a/.github/workflows/performance_tests.yml
+++ b/.github/workflows/performance_tests.yml
@@ -55,5 +55,5 @@ jobs:
         if: always()
         with:
           name: performance-plot-${{ github.run_number }}
-          path: ./python/tests/performance_plot.png
+          path: ./tests/performance_plot.png
           retention-days: 30