--- a/.github/workflows/pull_request.yml
+++ b/.github/workflows/pull_request.yml
@@ -38,13 +38,13 @@ jobs:
       E2B_TESTS_TEMPLATE: ${{ needs.build-template.outputs.template_id }}
   performance-tests:
     uses: ./.github/workflows/performance_tests.yml
-    # needs: build-template
+    needs: build-template
     secrets:
       E2B_API_KEY: ${{ secrets.E2B_API_KEY }}
     with:
       E2B_DOMAIN: ${{ vars.E2B_DOMAIN }}
-      # E2B_TESTS_TEMPLATE: ${{ needs.build-template.outputs.template_id }}
-      E2B_TESTS_TEMPLATE: "code-interpreter-v1"
+      E2B_TESTS_TEMPLATE: ${{ needs.build-template.outputs.template_id }}
+      E2B_TESTS_BENCHMARK_ITERATIONS_COUNT: 20
   cleanup-build-template:
     uses: ./.github/workflows/cleanup_build_template.yml
     needs: [build-template, js-sdk, python-sdk, performance-tests]