--- a/.github/workflows/pull_request.yml
+++ b/.github/workflows/pull_request.yml
@@ -44,7 +44,6 @@ jobs:
     with:
       E2B_DOMAIN: ${{ vars.E2B_DOMAIN }}
       E2B_TESTS_TEMPLATE: ${{ needs.build-template.outputs.template_id }}
-      E2B_TESTS_BENCHMARK_ITERATIONS_COUNT: 20
   cleanup-build-template:
     uses: ./.github/workflows/cleanup_build_template.yml
     needs: [build-template, js-sdk, python-sdk, performance-tests]