--- a/python/pyproject.toml
+++ b/python/pyproject.toml
@@ -24,6 +24,7 @@ pytest-asyncio = "^0.23.7"
 pytest-xdist = "^3.6.1"
 black = "23.12.1"
 pydoc-markdown = "^4.8.2"
+matplotlib = "^3.8.0"
 
 [build-system]
 requires = ["poetry-core"]